Decided even more what to do + moved the varables again
new plan once again: use mouse input to squash or stretch the sine and cosine waves because anything ivolving changing the trajectory of the waves while they were being drawn left me unsatified. also, made the variables global again because i'm going to need to use them in both setup() and draw() for this new plan. lastly, changed the size of the drawn ellipses to 5 pixels in diameter instead of 10